fix: 修复10项代码缺陷 — 逻辑错误、并发安全、安全隐患
- parse_tail 未匹配时返回空字符串而非原始文本
- close_all 将 _runtimes.clear() 移到清理循环之后防止资源泄漏
- MetadataStore 添加 asyncio.Lock 保护 SQLite 并发访问
- pickle.load 前检查文件所有权和权限防止任意代码执行
- active_memory 异常不再静默吞掉改为 logger.debug
- MemorixEvent.timestamp 类型从 int 改为 float 避免精度截断
- reinforce_buffer 从 Set 改为 Dict 实现确定性 FIFO 驱逐
- 删除 app_context.py 中重复的 _deep_merge 改为从 settings 导入
- 移除冗余的 atexit.register 依赖 FastAPI 生命周期清理
- injection system_rule 去重检查标准化空白防止重复注入

+++ memorix/adapters/astrbot_event_adapter.py
@@ -16,7 +16,7 @@ class MemorixEvent:
     group_id: str
     message_id: str
     message_text: str
-    timestamp: int
+    timestamp: float
 
 
 class AstrbotEventAdapter:
@@ -31,7 +31,7 @@ class AstrbotEventAdapter:
         group_id = str(getattr(event, "get_group_id", lambda: "")() or "")
         message_id = str(getattr(message_obj, "message_id", "") or "")
         message_text = str(getattr(event, "message_str", "") or "").strip()
-        timestamp = int(getattr(message_obj, "timestamp", 0) or 0)
+        timestamp = float(getattr(message_obj, "timestamp", 0) or 0)
         return MemorixEvent(
             scope_key=scope_key,
             platform=platform,
